limpeza cardProduto, cardRestaurant
