+++ src/Pages/InitialPage.js
@@ -44,9 +44,7 @@ export default function InitialPage() {
     const token = localStorage.getItem('token')
     
     setTimeout(()=>{
-      //usar o coordinator
       token? history.push('/feed') : history.push('login')
-      
     }, 2000)
 
   })
